@@ -1,149 +1,212 @@
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Net.Http;
 using System.Net.Http.Headers;
 using System.Text;
-using System.Threading.Tasks;
 using GitHubApiStatus;
 using YamlDotNet.Serialization;
 
-namespace AdvocateValidation
+namespace AdvocateValidation;
+class Program
 {
-    class Program
-    {
-        readonly static HttpClient _client = new();
-        readonly static GitHubApiStatusService _gitHubApiStatusService = new();
+    readonly static HttpClient _client = new();
+    readonly static GitHubApiStatusService _gitHubApiStatusService = new();
 
-        readonly static string _advocatesPath =
+    readonly static string _advocatesPath =
 #if DEBUG
-            Path.Combine("../../../../", "advocates");
+        Path.Combine("../../../../", "advocates");
 #else
             Path.Combine("../", "advocates");
 #endif
 
-        readonly static IDeserializer _yamlDeserializer = new DeserializerBuilder().Build();
+    readonly static IDeserializer _yamlDeserializer = new DeserializerBuilder().Build();
 
-        static Program()
+    static Program()
+    {
+        _client.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue(new ProductHeaderValue(nameof(AdvocateValidation))));
+    }
+
+    static async Task Main()
+    {
+        const string gitHub = "GitHub";
+
+        List<CloudAdvocateYamlModel> advocateList = [];
+
+        string[] advocateFiles = Directory.GetFiles(_advocatesPath);
+
+        List<(string, string)> parsingErrors = [];
+
+        await foreach ((string filePath, CloudAdvocateYamlModel advocate) in GetAdvocateYmlFiles(advocateFiles).ConfigureAwait(false))
         {
-            _client.DefaultRequestHeaders.UserAgent.Add(new ProductInfoHeaderValue(new ProductHeaderValue(nameof(AdvocateValidation))));
+            if (string.IsNullOrWhiteSpace(advocate?.Metadata.Alias))
+            {
+                parsingErrors.Add((filePath, "Missing Microsoft Alias"));
+                continue;
+            }
+
+            if (string.IsNullOrWhiteSpace(advocate.Metadata.Team))
+            {
+                parsingErrors.Add((filePath, "Missing Team"));
+                continue;
+            }
+
+            foreach (Connect connect in advocate.Connect)
+            {
+                if (connect.Title.Equals("LinkedIn", StringComparison.OrdinalIgnoreCase))
+                {
+#if DEBUG
+                    Console.WriteLine($"LinkedIn doesn't like being validated, it returns 999 response codes, skipping {connect.Url} from {filePath}.");
+#endif
+                    continue;
+                }
+
+                try
+                {
+                    if (connect.Title.Equals(gitHub, StringComparison.OrdinalIgnoreCase))
+                    {
+                        await EnsureValidGitHubUri(filePath, connect.Url, gitHub).ConfigureAwait(false);
+                    }
+                    else
+                    {
+                        await EnsureValidUri(filePath, connect.Url, connect.Title).ConfigureAwait(false);
+                    }
+                }
+                catch (Exception ex)
+                {
+                    parsingErrors.Add((filePath, ex.Message));
+                }
+            }
+
+            try
+            {
+                EnsureValidImage(filePath, advocate.Image);
+            }
+            catch (Exception ex)
+            {
+                parsingErrors.Add((filePath, ex.Message));
+            }
+
+            advocateList.Add(advocate);
         }
 
-        static async Task Main()
+        IEnumerable<string> duplicateAliasList = advocateList.GroupBy(x => x.Metadata.Alias).Where(g => g.Count() > 1).Select(x => x.Key);
+        foreach (string duplicateAlias in duplicateAliasList)
         {
-            const string gitHub = "GitHub";
+            parsingErrors.Add(("", $"Duplicate Alias Found; ms.author: {duplicateAlias}"));
+        }
 
-            var advocateList = new List<CloudAdvocateYamlModel>();
-
-            var advocateFiles = Directory.GetFiles(_advocatesPath);
-
-            await foreach (var (filePath, advocate) in GetAdvocateYmlFiles(advocateFiles).ConfigureAwait(false))
+        if (parsingErrors.Count != 0)
+        {
+            Console.WriteLine("Validation Failed");
+            foreach ((string filePath, string parsingError) in parsingErrors)
             {
-                if (string.IsNullOrWhiteSpace(advocate?.Metadata.Alias))
-                    throw new Exception($"Missing Microsoft Alias: {filePath}");
-
-                if (string.IsNullOrWhiteSpace(advocate.Metadata.Team))
-                    throw new Exception($"Missing Team: {filePath}");
-
-                var gitHubUri = advocate.Connect.FirstOrDefault(x => x.Title.Equals(gitHub, StringComparison.OrdinalIgnoreCase))?.Url;
-
-                await EnsureValidGitHubUri(filePath, gitHubUri, gitHub).ConfigureAwait(false);
-
-                EnsureValidImage(filePath, advocate.Image);
-
-                advocateList.Add(advocate);
+                Console.WriteLine($"::error file={filePath}::{parsingError}");
             }
-
-            var duplicateAliasList = advocateList.GroupBy(x => x.Metadata.Alias).Where(g => g.Count() > 1).Select(x => x.Key);
-            foreach (var duplicateAlias in duplicateAliasList)
-            {
-                throw new Exception($"Duplicate Alias Found; ms.author: {duplicateAlias}");
-            }
-
+            throw new Exception("Validation Failed");
+        }
+        else
+        {
             Console.WriteLine("Validation Completed Successfully");
         }
+    }
 
-        static async IAsyncEnumerable<(string filePath, CloudAdvocateYamlModel advocate)> GetAdvocateYmlFiles(IEnumerable<string> files)
+    static async IAsyncEnumerable<(string filePath, CloudAdvocateYamlModel advocate)> GetAdvocateYmlFiles(IEnumerable<string> files)
+    {
+        var ymlFiles = files.Where(x => x.EndsWith(".yml", StringComparison.OrdinalIgnoreCase));
+
+        foreach (var filePath in ymlFiles)
         {
-            var ymlFiles = files.Where(x => x.EndsWith(".yml", StringComparison.OrdinalIgnoreCase));
+            var text = await File.ReadAllTextAsync(filePath).ConfigureAwait(false);
 
-            foreach (var filePath in ymlFiles)
+            if (text.StartsWith("### YamlMime:Profile") && !text.StartsWith("### YamlMime:ProfileList"))
             {
-                var text = await File.ReadAllTextAsync(filePath).ConfigureAwait(false);
-
-                if (text.Contains($"{Uri.UriSchemeHttp}://"))
-                    throw new Exception($"Invalid Uri. Must Use HTTPS: {filePath}");
-
-                if (text.StartsWith("### YamlMime:Profile") && !text.StartsWith("### YamlMime:ProfileList"))
-                {
-                    Console.WriteLine($"Parsing {filePath}");
-                    yield return (filePath, ParseAdvocateFromYaml(text));
-                }
+                Console.WriteLine($"Parsing {filePath}");
+                yield return (filePath, ParseAdvocateFromYaml(text));
             }
         }
+    }
 
-        static CloudAdvocateYamlModel ParseAdvocateFromYaml(in string fileText)
+    static CloudAdvocateYamlModel ParseAdvocateFromYaml(in string fileText)
+    {
+        var stringReaderFile = new StringReader(fileText);
+
+        return _yamlDeserializer.Deserialize<CloudAdvocateYamlModel>(stringReaderFile);
+    }
+
+    static async Task EnsureValidUri(string filePath, Uri? uri, string uriName)
+    {
+        if (uri is null)
+            throw new Exception($"Missing {uriName} Url: {uri}, File: {filePath}");
+
+        if (!uri.IsWellFormedOriginalString())
+            throw new Exception($"URI for {uriName} is malformed. Url: {uri}, File: {filePath}");
+
+        if (uri.Scheme == Uri.UriSchemeHttp)
+            Console.WriteLine($"::warning file={filePath}:: {uriName} Url is HTTP, you really should be hosting on HTTPS. Url: {uri}.");
+
+        try
         {
-            var stringReaderFile = new StringReader(fileText);
+            HttpResponseMessage response = await _client.GetAsync(uri).ConfigureAwait(false);
 
-            return _yamlDeserializer.Deserialize<CloudAdvocateYamlModel>(stringReaderFile);
+            if (!response.IsSuccessStatusCode)
+                throw new Exception($"Failed to resolve URI for {uriName}. Url: {uri}, File: {filePath}");
         }
-
-        static async Task EnsureValidGitHubUri(string filePath, Uri? uri, string uriName)
+        catch (Exception ex)
         {
-            if (uri is null)
-                throw new Exception($"Missing {uriName} Url: {filePath}");
+            throw new Exception($"Failed to resolve URI for {uriName}. Url: {uri}, File: {filePath}", ex);
+        }
+    }
 
-            if (!uri.IsWellFormedOriginalString())
-                throw new Exception($"Invalid {uriName} Url: {filePath}");
+    static async Task EnsureValidGitHubUri(string filePath, Uri? uri, string uriName)
+    {
+        if (uri is null)
+            throw new Exception($"Missing {uriName} Url: {uri}, File: {filePath}");
 
-            bool hasReceivedGitHubAbuseLimitResponse;
+        if (!uri.IsWellFormedOriginalString())
+            throw new Exception($"Invalid {uriName} Url: {uri}, File: {filePath}");
 
-            do
+        bool hasReceivedGitHubAbuseLimitResponse;
+
+        do
+        {
+            HttpResponseMessage response = await _client.GetAsync(uri).ConfigureAwait(false);
+            hasReceivedGitHubAbuseLimitResponse = _gitHubApiStatusService.IsAbuseRateLimit(response.Headers, out var delta);
+
+            if (hasReceivedGitHubAbuseLimitResponse && delta is TimeSpan timeRemaining)
             {
-                var response = await _client.GetAsync(uri).ConfigureAwait(false);
-                hasReceivedGitHubAbuseLimitResponse = _gitHubApiStatusService.IsAbuseRateLimit(response.Headers, out var delta);
-
-                if (hasReceivedGitHubAbuseLimitResponse && delta is TimeSpan timeRemaining)
-                {
-                    Console.WriteLine($"Rate Limit Exceeded. Retrying in {timeRemaining.TotalSeconds} seconds");
-                    await Task.Delay(timeRemaining).ConfigureAwait(false);
-                }
-                else if (!response.IsSuccessStatusCode)
-                {
-                    throw new Exception($"Invalid {uriName} Url: {filePath}");
-                }
+                Console.WriteLine($"Rate Limit Exceeded. Retrying in {timeRemaining.TotalSeconds} seconds");
+                await Task.Delay(timeRemaining).ConfigureAwait(false);
             }
-            while (hasReceivedGitHubAbuseLimitResponse);
+            else if (!response.IsSuccessStatusCode)
+            {
+                throw new Exception($"Invalid {uriName} Url: {uri}, File: {filePath}");
+            }
         }
+        while (hasReceivedGitHubAbuseLimitResponse);
+    }
 
-        static void EnsureValidImage(in string filePath, in Image? cloudAdvocateImage)
-        {
-            if (cloudAdvocateImage is null)
-                throw new Exception($"Image Source Missing: {filePath}");
+    static void EnsureValidImage(in string filePath, in Image? cloudAdvocateImage)
+    {
+        if (cloudAdvocateImage is null)
+            throw new Exception($"Image Source Missing: {filePath}");
 
-            if (string.IsNullOrWhiteSpace(cloudAdvocateImage.Src))
-                throw new Exception($"Image Source Missing: {filePath}");
+        if (string.IsNullOrWhiteSpace(cloudAdvocateImage.Src))
+            throw new Exception($"Image Source Missing: {filePath}");
 
-            var filePathRelativeToValidation = Path.Combine(_advocatesPath, cloudAdvocateImage.Src);
+        string filePathRelativeToValidation = Path.Combine(_advocatesPath, cloudAdvocateImage.Src);
 
-            var fileStream = new FileStream(filePathRelativeToValidation, FileMode.Open);
-            var binaryReader = new BinaryReader(fileStream, Encoding.UTF8);
+        using FileStream fileStream = new(filePathRelativeToValidation, FileMode.Open);
+        using BinaryReader binaryReader = new(fileStream, Encoding.UTF8);
 
-            var imageSize = ImageService.GetDimensions(binaryReader);
+        System.Drawing.Size imageSize = ImageService.GetDimensions(binaryReader);
 
-            if (imageSize.Height <= 0)
-                throw new Exception($"Invalid Image Height (must be greater than 0): {filePath}");
+        if (imageSize.Height <= 0)
+            throw new Exception($"Invalid Image Height (must be greater than 0): {filePath}");
 
-            if (imageSize.Width <= 0)
-                throw new Exception($"Invalid Image Width (must be greater than 0): {filePath}");
+        if (imageSize.Width <= 0)
+            throw new Exception($"Invalid Image Width (must be greater than 0): {filePath}");
 
-            if (imageSize.Height != imageSize.Width)
-                throw new Exception($"Invalid Image (Height and Width must be equal): {filePath}");
+        if (imageSize.Height != imageSize.Width)
+            throw new Exception($"Invalid Image (Height and Width must be equal): {filePath}");
 
-            if (cloudAdvocateImage.Alt is null)
-                throw new Exception($"Image Alt Text Missing: {filePath}");
-        }
+        if (cloudAdvocateImage.Alt is null)
+            throw new Exception($"Image Alt Text Missing: {filePath}");
     }
 }
